# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: ecom application
  title: swagger documentation for the application
  version: "1.0"
paths:
...
  /user/profile:
    get:
      description: Get user details
      produces:
      - application/json
      responses:
        "200":
          description: listed Profile Details
          schema:
            type: json
        "401":
          description: Failed to list Profile Details
          schema:
            type: json
      summary: Get user details
      tags:
      - User-Profile
    patch:
      description: Edit user details
      produces:
      - application/json
      responses:
        "200":
          description: Edited User Details
          schema:
            type: json
        "401":
          description: Failed to edit User Details
          schema:
            type: json
      summary: Edit user details
      tags:
      - User-Profile
...
swagger: "2.0"